// mock/src/generated/prisma/enums.ts
/* !!! This is code generated by Prisma. Do not edit directly. !!! */
/* eslint-disable */
// biome-ignore-all lint: generated file
// @ts-nocheck 
/*
* This file exports all enum related types from the schema.
*
* 🟢 You can import this file directly.
*/

export const Plataforma = {
  IFOOD: 'IFOOD',
  NOVENTA_NOVE_FOOD: 'NOVENTA_NOVE_FOOD',
  KEETA: 'KEETA',
  UBER_EATS: 'UBER_EATS',
  RAPPI: 'RAPPI'
} as const

export type Plataforma = (typeof Plataforma)[keyof typeof Plataforma]


export const TipoPedido = {
  ENTREGA: 'ENTREGA',
  RETIRADA: 'RETIRADA'
} as const

export type TipoPedido = (typeof TipoPedido)[keyof typeof TipoPedido]


export const EntreguePor = {
  PLATAFORMA: 'PLATAFORMA',
  LOJA: 'LOJA'
} as const

export type EntreguePor = (typeof EntreguePor)[keyof typeof EntreguePor]


export const CanalEvento = {
  WEBHOOK: 'WEBHOOK',
  POLLING: 'POLLING'
} as const

export type CanalEvento = (typeof CanalEvento)[keyof typeof CanalEvento]